pyproject.toml
[tool.poetry]
name = "pp2-util-bot"
version = "0.1.1"
description = "Utility-service Telegram Bot for Parque dos Passaros 2 Residential Condominium"
authors = ["Kevin Mian Kraiker <kevin.mian@gmail.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"
packages = [{include = "pp2_util_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.6"
environs = "^9.5.0"
nullsafe = "^0.2.1"

[tool.poetry.group.dev.dependencies]
flakeheaven = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flakeheaven]
base = "https://raw.githubusercontent.com/flakeheaven/flakeheaven/master/pyproject.toml"
max_line_length = 120
show_source = true

[tool.flakeheaven.plugins]
pycodestyle = ["+*", "-W503"]
